design: 쿠폰 반응형 flex 적용
